+++ cs/Markdown/Helpers/TagWithIndex.cs
@@ -1,0 +1,15 @@
+using Markdown.Tags;
+
+namespace Markdown.Helpers;
+
+public class TagWithIndex
+{
+    public ITag? Tag;
+    public int Index;
+
+    public TagWithIndex(ITag? tag, int index)
+    {
+        Tag = tag;
+        Index = index;
+    }
+}
